getting the packaging setup
