+++ requirements.txt
@@ -1,0 +1,49 @@
+addict==2.4.0
+alembic==1.11.1
+argcomplete==1.10.3
+beautifulsoup4==4.8.2
+certifi==2023.7.22
+chardet==3.0.4
+charset-normalizer==3.3.1
+click==8.1.7
+compressed-rtf==1.0.6
+dill==0.3.7
+docx2txt==0.8
+ebcdic==1.1.1
+extract-msg==0.28.7
+greenlet==3.0.0
+idna==3.7
+IMAPClient==2.1.0
+imbalanced-learn==0.11.0
+joblib==1.3.2
+lxml==4.9.3
+nltk==3.8.1
+numpy==1.26.1
+olefile==0.46
+pdfminer.six==20191110
+Pillow==10.2.0
+psycopg2-binary==2.9.9
+pycryptodome==3.19.1
+python-dateutil==2.8.2
+python-json-logger==2.0.7
+python-pptx==0.6.22
+PyYAML==6.0.1
+regex==2023.10.3
+requests==2.31.0
+scikit-learn==1.2.2
+scipy==1.11.3
+six==1.12.0
+sortedcontainers==2.4.0
+soupsieve==2.5
+SpeechRecognition==3.8.1
+SQLAlchemy==2.0.22
+SQLAlchemy-Utils==0.41.1
+textract==1.6.5
+threadpoolctl==3.2.0
+tqdm==4.66.1
+typing_extensions==4.8.0
+tzlocal==5.2
+urllib3==2.0.7
+wget==3.2
+xlrd==1.2.0
+XlsxWriter==3.1.9
