+++ setup.cfg
@@ -1,0 +1,63 @@
+[metadata]
+name = srt-ml
+version = 0.0.1
+author = Adam Buckingham
+author_email = adam.buckingham@gsa.gov
+description = Machine Learning Algorithm development for the Solicitation Review Tool
+license = CC0 1.0 Universal
+platforms = unix, linux, osx
+classifiers =
+    Programming Language :: Python :: 3 
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.10
+
+[options]
+packages = find:
+install_requires = 
+    addict
+    dill
+    imbalanced-learn
+    nltk
+    numpy
+    pandas
+    psycopg2-binary
+    python-dateutil
+    PyYAML
+    scikit-learn==1.2.2
+    scipy
+    textract
+python_requires = >=3.10
+package_dir =
+    =src
+zip_safe = False
+include_package_data = True
+
+
+[options.extras_require]
+dev =
+    alembic
+    fpdf
+    python-docx
+    pytest
+    pytest-cov
+    mypy
+    requests_mock
+    reportlab
+    ruff
+
+[options.package_data]
+mypkg =
+    *.pkl
+
+
+[options.packages.find]
+where=src
+exclude =
+    cf*
+    bin*
+    docs*
+
+
+[options.entry_points]
+
+
