+++ src/srt_ml/__init__.py
@@ -1,0 +1,2 @@
+from .binaries import binary_path
+from .predict.predict import Predict
