+++ src/srt_ml/binaries/__init__.py
@@ -1,0 +1,2 @@
+from pathlib import Path
+binary_path = Path(__file__).parent
