--- srt-ml/binaries/__init__.py
@@ -1,2 +1,0 @@
-from pathlib import Path
-binary_path = Path(__file__).parent
